feat(streaming): add proto for lookup executor
