+++ src/compute/tests/table_v2_materialize.rs
@@ -173,7 +173,6 @@ async fn test_table_v2_materialize() -> Result<()> {
         all_column_ids.clone(),
         2,
         "MaterializeExecutor".to_string(),
-        vec![],
     ))
     .v1();
