+++ src/stream/src/executor_v2/lookup.rs
@@ -15,14 +15,24 @@
 use async_trait::async_trait;
 use futures::StreamExt;
 use risingwave_common::catalog::Schema;
+use risingwave_common::error::Result;
+use risingwave_common::try_match_expand;
 use risingwave_common::types::DataType;
-use risingwave_storage::StateStore;
+use risingwave_pb::stream_plan;
+use risingwave_pb::stream_plan::stream_node::Node;
+use risingwave_storage::{Keyspace, StateStore};
 
+use crate::executor::{Executor as ExecutorV1, ExecutorBuilder};
 use crate::executor_v2::{Barrier, BoxedMessageStream, Executor, PkIndices, PkIndicesRef};
+use crate::task::{ExecutorParams, LocalStreamManagerCore};
 
 mod sides;
 use self::sides::*;
 mod impl_;
+
+pub use impl_::LookupExecutorParams;
+
+use super::ExecutorV1AsV2;
 
 #[cfg(test)]
 mod tests;
@@ -74,5 +84,38 @@ impl<S: StateStore> Executor for LookupExecutor<S> {
 
     fn identity(&self) -> &str {
         "<unknown>"
+    }
+}
+
+pub struct LookupExecutorBuilder {}
+
+impl ExecutorBuilder for LookupExecutorBuilder {
+    fn new_boxed_executor(
+        mut params: ExecutorParams,
+        node: &stream_plan::StreamNode,
+        store: impl StateStore,
+        _stream: &mut LocalStreamManagerCore,
+    ) -> Result<Box<dyn ExecutorV1>> {
+        let node = try_match_expand!(node.get_node().unwrap(), Node::LookupNode)?;
+
+        let stream = params.input.remove(1);
+        let stream = Box::new(ExecutorV1AsV2(stream));
+        let arrangement = params.input.remove(0);
+        let arrangement = Box::new(ExecutorV1AsV2(arrangement));
+
+        Ok(Box::new(
+            Box::new(LookupExecutor::new(LookupExecutorParams {
+                arrangement,
+                stream,
+                arrangement_keyspace: Keyspace::shared_executor_root(store, u64::MAX),
+                arrangement_col_descs: vec![], // TODO: fill this field
+                arrangement_order_rules: vec![], // TODO: fill this field
+                pk_indices: params.pk_indices,
+                use_current_epoch: node.use_current_epoch,
+                stream_join_key_indices: node.stream_key.iter().map(|x| *x as usize).collect(),
+                arrange_join_key_indices: node.arrange_key.iter().map(|x| *x as usize).collect(),
+            }))
+            .v1(),
+        ))
     }
 }
