+++ src/stream/src/executor_v2/lookup/tests.rs
@@ -128,7 +128,6 @@ async fn create_arrangement(
         arrangement_col_arrange_rules(),
         column_ids,
         1,
-        vec![],
     ))
 }
